chore: Update dependencies and features
- Update `smithay/smithay` to `f208cd7`
    > All changes to code are to support smithay changes

- Fix irrefutable_let_patterns occuring when building with `x11_backend`
  or `udev_backend` feature exclusively.

- Run `cargo update`
    > egui still at 0.23, waiting for smithay to update to `glow` v0.13

- Make `zbus` and `zvariant` optional

- Get rid of `nferhat/iced` as a workspace dependency
    > Was needed for `fht-share-picker` but got removed

+++ src/shell/grabs.rs
@@ -49,7 +49,7 @@ impl PointerGrab<State> for MoveSurfaceGrab {
         &mut self,
         data: &mut State,
         handle: &mut PointerInnerHandle<'_, State>,
-        _: Option<(PointerFocusTarget, Point<i32, Logical>)>,
+        _: Option<(PointerFocusTarget, Point<f64, Logical>)>,
         event: &MotionEvent,
     ) {
         // While the grab handle is active, no client should have focus, otherwise bad stuff WILL
@@ -99,7 +99,7 @@ impl PointerGrab<State> for MoveSurfaceGrab {
         &mut self,
         data: &mut State,
         handle: &mut PointerInnerHandle<'_, State>,
-        focus: Option<(PointerFocusTarget, Point<i32, Logical>)>,
+        focus: Option<(PointerFocusTarget, Point<f64, Logical>)>,
         event: &RelativeMotionEvent,
     ) {
         handle.relative_motion(data, focus, event);
